Just fixing file system and almost ready

+++ EXPERIMENTS/bin_eval_per_obj_type_ann_norm_small_med/objects/recycler_obj.py
@@ -1,11 +1,13 @@
 import os
 import shutil
+import logging
 
 class recyclerObj:
-    def __init__(self, prev_trial_folder, current_folder, files_to_recycle):
+    def __init__(self, prev_trial_folder, current_folder, files_to_recycle, current_trial_number):
         self.prev_trial_folder = prev_trial_folder
         self.current_folder = current_folder
         self.files_to_recycle = files_to_recycle
+        self.current_trial_number = current_trial_number
 
     def copy_subfolders(self):
         for subdir, _, files in os.walk(self.prev_trial_folder):
@@ -18,4 +20,8 @@ class recyclerObj:
                 if f in files:
                     src_file = os.path.join(subdir, f)
                     dst_file = os.path.join(new_subdir, f)
-                    shutil.copy(src_file, dst_file)
+                    if not os.path.exists(dst_file):
+                        logging.info(f"Recycling file {src_file}")
+                        shutil.copy(src_file, dst_file)
+                    else:
+                        logging.info(f"File {src_file} already present in Trial #{self.current_trial_number}, not recycling ...")
